api/pyproject.toml
[project]
name = "eccc-hudson-bay-lowlands-api"
version = "0.1.0"
description = "ECCC Hudson Bay Lowlands backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]",
    "titiler-core==1.1.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.0",
    "shapely>=2.0",
    "pyproj>=3.6",
    "geoalchemy2>=0.15",
]

[tool.setuptools]
packages = ["db", "models", "routers", "schemas", "services"]

[tool.uv]
dev-dependencies = [
    "pytest",
    "httpx",
    "ruff",
]

[tool.ruff]
line-length = 127

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
